requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==3.7.0
django-crispy-forms==2.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg==3.1.13
psycopg-binary==3.1.13
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
